Narrow pytest fixture scope to fix memory issues 2.8.x (#9687)
