+++ tests/conftest.py
@@ -284,16 +284,16 @@ async def nlu_agent(trained_nlu_model: Text) -> Agent:
     return await load_agent(model_path=trained_nlu_model)
 
 
-@pytest.fixture(scope="session")
+@pytest.fixture(scope="module")
 async def unexpected_intent_policy_agent(
     trained_unexpected_intent_policy_path: Text,
 ) -> Agent:
     return await load_agent(model_path=trained_unexpected_intent_policy_path)
 
 
-@pytest.fixture(scope="session")
-async def mood_agent(trained_moodbot_path: Text) -> Agent:
-    return await load_agent(model_path=trained_moodbot_path)
+@pytest.fixture(scope="module")
+def mood_agent(trained_moodbot_path: Text) -> Agent:
+    return Agent.load_local_model(model_path=trained_moodbot_path)
 
 
 @pytest.fixture(scope="session")
@@ -524,24 +524,6 @@ async def trained_response_selector_bot(trained_async: Callable) -> Path:
 
 
 @pytest.fixture(scope="session")
-async def trained_restaurantbot(trained_async: Callable) -> Path:
-    zipped_model = await trained_async(
-        domain="data/test_restaurantbot/domain.yml",
-        config="data/test_restaurantbot/config.yml",
-        training_files=[
-            "data/test_restaurantbot/data/rules.yml",
-            "data/test_restaurantbot/data/stories.yml",
-            "data/test_restaurantbot/data/nlu.yml",
-        ],
-    )
-
-    if not zipped_model:
-        raise RasaException("Model training for formbot failed.")
-
-    return Path(zipped_model)
-
-
-@pytest.fixture(scope="session")
 async def e2e_bot_domain_file() -> Path:
     return Path("data/test_e2ebot/domain.yml")
 
@@ -584,24 +566,19 @@ async def e2e_bot(
     return Path(zipped_model)
 
 
-@pytest.fixture(scope="session")
+@pytest.fixture(scope="module")
 async def response_selector_agent(
     trained_response_selector_bot: Optional[Path],
 ) -> Agent:
     return Agent.load_local_model(trained_response_selector_bot)
 
 
-@pytest.fixture(scope="session")
-async def restaurantbot_agent(trained_restaurantbot: Optional[Path],) -> Agent:
-    return Agent.load_local_model(trained_restaurantbot)
-
-
-@pytest.fixture(scope="session")
+@pytest.fixture(scope="module")
 async def response_selector_interpreter(response_selector_agent: Agent,) -> Interpreter:
     return response_selector_agent.interpreter.interpreter
 
 
-@pytest.fixture(scope="session")
+@pytest.fixture(scope="module")
 async def e2e_bot_agent(e2e_bot: Optional[Path],) -> Agent:
     return Agent.load_local_model(e2e_bot)
